fw-api: Update HW header files to version R102 for QCA6290
Update HW include header files to DV release version R102
from R81. This is required to support emulation
version E3 for all platforms. Note that this change only
updates the _EXISTING_ header files, which already existed
in the earlier version, and is also present in R102. Files
which are required to be newly added to the git repository
from the HW repository, as part of R102 upgrade will be
added using a separate change.

Change-Id: I934a343998d965457d2c5ffbc7dee31ec5208935
CRs-Fixed: 1105853