qcacld-3.0: Sanitize the sme_get_link_speed() API
Currently sme_get_link_speed() requires that the caller pass an
allocated buffer as the lsReq param. This violates the general
concensus that SME should make no assumptions about the buffers passed
to it by HDD.

Update the implementation such that SME allocates the buffer that will
be passed in the message, and update HDD to use the stack.

Note that this is a prerequisite for moving the linkspeed request
processing to the new HDD Request Management infrastructure.

Change-Id: I09257855c6c260bb4a6ca91eb79bb24813cc5efa
CRs-Fixed: 2000023