qcacmn: DFS componentization (NOL)
NOL stands for non-occupancy list. After RADAR has been detected in a
particular channel,  the channel cannot be used for a period of 30 minutes,
which is called the non-occupancy. The NOL is basically a list of all the
channels on which RADAR has been detected. Each channel has a 30-minute
timer associated with it.

Change-Id: Ie823ef86a4109ff200c9bb2307711d17ff3f9876
CRs-Fixed: 2001106