qcacmn: Remove DPU signatures from TDLS Add STA Response
The ucast_sig and bcast_sig fields in struct tdls_add_sta_rsp date
back to a previous version of the driver that used these signatures to
synchronize the data plane with the control plane. However the current
version of the driver does not use these parameters, so remove them as
the final stage of the plan to completely remove these signatures from
throughout the driver.

Since the converged struct tdls_add_sta_rsp must exactly match the
legacy tSirTdlsAddStaRsp data structure, this change is co-dependent
with the qcacld-3.0 change:
  qcacld-3.0: Remove DPU signatures from TDLS Add STA Response
  Change-Id: I765b090fc5dbe9bdaf83b05a72223c0a12f71a18

Change-Id: I7750fac9e4870d75358606bea427a5dbab97c665
CRs-Fixed: 2200973