asoc: msm-pcm-routing: add Listen app type config v2 mixer control
It is required to know the number of output channels for any
particular app_type so that channel mixing can be performed
if the channels do not match as that of the listen stream.
Add new Listen app type config V2 mixer control that reads the
num_out_channels field along with the others.

Change-Id: I49a74004e12dfa5a4c1ebaac4ac065acc6ff5bc3
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>