qcacld-3.0: Pass correct pointer type to mac_open()
In cds_open() the call to mac_open() currently passes a pointer to the
gp_cds_context->mac_context field. This is a void * field but
mac_open() expects a mac_handle_t * (aka tHalHandle *). This currently
compiles because mac_handle_t is itself a void *. But in the future
the underlying type of mac_handle_t will change, so pass a pointer to
an object of the correct type to mac_open().

Change-Id: Ie3e54c5288bc7d1e9287f30d997b005a5c8dc2ef
CRs-Fixed: 2268813