qcacld-3.0: Don't stop transition in ssr shutdown
During sub-system recovery, the platform driver invokes the wlan's
shutdown and reinit callbacks. Reinit is called regardless of the result
of shutdown. Currently, the SSR transition is aborted in shutdown in the
event of a failure. This results in a panic during reinit, which expects
to be protected by that transition. Regardless of the result of
shutdown, do not stop the SSR transition such that the entire recovery
process is properly protected.

Change-Id: I3685920de5a7d47f6334bf4d8389cb267a46fac2
CRs-Fixed: 2350868