qcacld-3.0: Replace typedef tSirEseBcnReportRsp
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirEseBcnReportRsp typedef
does not meet any of those criteria, so replace it (and the "tp"
variant) with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names
and so-called Hungarian notation, so in conjunction rename the
underlying struct to be in compliance.

Change-Id: Id687d039c51eaf293cdd39536cf200874bd8fbc9
CRs-Fixed: 2394253