disp: msm: sde: delay backlight update until the first commit
In the current implementation panel set backlight can be called even before
receiving a frame from the driver during the first commit. Hence there is
chance for some garbage content to be shown onto the panel. So this change
imposes a condition to delay backlight update of the panel until the
first frame is received from the HW.

This change includes only a missing portion of the original that was somehow
left out of the snapshot ported to 4.19. Original change ID is preserved.

Change-Id: I735a97088efef91159e724bba615549dba49e4e3
Signed-off-by: Tharun Raj Soma <tsoma@codeaurora.org>
Signed-off-by: Steve Cohen <cohens@codeaurora.org>