qcacmn: Filter the channel list according to acs cfg
Currently the driver checks whether the DFS channel is
in the range of acs channel list given, i.e the start
and end channel, which is not always correct as the channel
range does not imply that all channels in that range are present.

Fix is to explicitly check for the number of channels, and
compare each channel in the acs channel list with the DFS
channel.

Change-Id: Ib17c5e549a2c62652a6333d5d13ca8ff5a7b1674
CRs-Fixed: 2466228