qcacld-3.0: Change check for wma_is_vdev_started() to WLAN_VDEV_S_UP
When running WHUNT test case for SET NUD STATS vendor command,
wma_set_arp_req_stats is supposed to be sent to FW but it fails to send
due to the check in wma_is_vdev_started(), which is currently checking
if the vdev state is WLAN_VDEV_S_START, instead of WLAN_VDEV_S_UP, thus
allowing vdev state WLAN_VDEV_S_UP to still continue when it is not
started. Hence, change the check to WLAN_VDEV_S_UP to provide proper check.

Change-Id: Ifae00ff26a27df92cd4d09047c326bcccfdbf602
CRs-Fixed: 2651406