bindings: Documentation: add a property to exclude reg-names
In trusted vm, register ranges that need to be mapped need to exactly be
the same as that of HLOS vm. But the DT provided register ranges are used
to validate the IO memory assign from HLOS vm. So this change introduces a
separate property to specify the register names which need to be excluded
from the IO memory validation list but still need to be mapped to the vm.

Change-Id: I8c5c194254ae260f7a8878bb5492f5a039203fc9