qcacld-3.0: Don't connect in 11ax if peer ldpc is disabled
Don't connect in HE mode if the AP operating in 2.4 GHz does not
support LDPC coding.
Also, some peers which operate on 2.4 ghz and don't support LDPC
coding, might satisfy the condtions which is used to force enable
ldpc coding bit. If this bit is set, the peer doesn't support ldpc,
it  might fail to connect.
To fix this, force enable ldpc coding only when peer is operating
in 5/6 Ghz along with the already existing conditions.

Change-Id: I4b424833bbf845de17acc0907e1098c0a01b3ee5
CRs-Fixed: 2760506