msm: camera: isp: Change csid stop sequence during flush
This change takes care of a race which we see when we halt master
csid immediately and slave csid halt to follow. In some scenarios
after master csid halt, slave gives violation errors.
With this change we are first changing the halt mode of slave to
internal so that dependency on master is removed and
then issue halt commands to both master and slave.

CRs-Fixed: 2785781
Change-Id: Ib84a8c416e05a4ac54b6afad5c06017931d67a05
Signed-off-by: Shravya Samala <shravyas@codeaurora.org>