ARM: dts: msm: add MX entries to qcom,bw-scale for PCIe on lahaina
Explicit MX voting and scaling is required for PCIe on some
chipsets and there are updates to PCIe bus driver to support
this. Although PCIe on lahaina does not need to explicitly vote
for MX, updates to lahaina PCIe DT nodes are required to be
compatible with the new qcom,bw-scale structure. Therefore, add
MX entries to qcom,bw-scale for both PCIe nodes on lahaina.

Change-Id: If707f61f78e4d99df118ebbfb516d639e5b4f851