qcacld-3.0: Close/destroy the pdev before CDS closing
Pdev is created once all the modules are enabled and currently
it is freed after the cds close is done, if some of the modules
as part of the de-init sequence has to send a command to the firmware
it will result in error.

To keep the deinit sequence opposite to that of init close/destroy
the pdev before closing the cds.

Change-Id: I8e087a573112e5d2d81e8ea413044d45f228813e
CRs-Fixed: 2797344