pinctrl: distinguish reserve gpio list by macro definition
For each project has its own reserve gpio list. So we should have
the way to distinguish it in every different products.
We use a macro definition for this. If the gpios are not assigned to
TZ, they should be removed from the list.

Change-Id: I0837a3abdf8d7398aa2e7138563cb891b9f53780
Reviewed-on: https://gerrit.mot.com/2121376
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key